requirements.txt
openai==0.28.0
tweepy==4.14.0
python-dotenv==1.0.0